added exact_plus option/filter for RM counting

+++ rm_by_sample.py
@@ -91,8 +91,11 @@ def getNumPosArr():
 	return np_list
 
 
-def writeRMInfoToTable(dir_base,out_tbl):
+def writeRMInfoToTable(dir_base,out_tbl,num_rm_filter_thresh="0+"):
 	file_num=0
+	validation_res=validate_filter_stat_arg(num_rm_filter_thresh)
+	thresh_num=validation_res[0] #a non-negative integer
+	num_plus=validation_res[1] #True or False
 	for root, dirs, files in sorted(os.walk(dir_base)):
 		for items in sorted(fnmatch.filter(files, "*out.tsv")):
 			cdr1_lens=[0,0,0]
@@ -114,19 +117,25 @@ def writeRMInfoToTable(dir_base,out_tbl):
 						cdr1_amino_len=CDR1_len/3
 						cdr1_lens[cdr1_amino_len-5]+=1
 						rms=eval(pieces[185])
-						for rm in rms:
-							rm_str_len=len(rm)
-							if(rm_str_len==4):
-								numbered_pos=rm[1:3]
-							elif(rm_str_len==5):
-								numbered_pos=rm[1:4]
-							else:
-								raise Exception("Error, unknown number (extraction pos) in  numbered RM : "+rm)
-							if(numbered_pos in rm_count_map):
-								rm_count_map[numbered_pos]+=1
-							else:
-								raise Exception("Error, unseen number (extraction pos) in  numbered RM : "+rm)
-								
+						num_rms_in_this_row=len(rms)
+						examine_counts_for_this_row=False
+						if(num_plus and num_rms_in_this_row>=thresh_num):
+							examine_counts_for_this_row=True
+						if(not(num_plus) and num_rms_in_this_row==thresh_num):
+							examine_counts_for_this_row=True
+						if(examine_counts_for_this_row):
+							for rm in rms:
+								rm_str_len=len(rm)
+								if(rm_str_len==4):
+									numbered_pos=rm[1:3]
+								elif(rm_str_len==5):
+									numbered_pos=rm[1:4]
+								else:
+									raise Exception("Error, unknown number (extraction pos) in  numbered RM : "+rm)
+								if(numbered_pos in rm_count_map):
+									rm_count_map[numbered_pos]+=1
+								else:
+									raise Exception("Error, unseen number (extraction pos) in  numbered RM : "+rm)
 				line_num+=1
 			reader.close()
 			nPos_arr=getNumPosArr()
@@ -155,20 +164,48 @@ def writeRMInfoToTable(dir_base,out_tbl):
 			
 
 
+def validate_filter_stat_arg(v):
+	import re
+	valid_re=re.compile(r'^(\d+)(\+?)$')
+	search_res=re.search(valid_re,v)
+	if(search_res):
+		num=int(search_res.group(1))
+		plus=search_res.group(2)
+		print "num=",num," plus=",plus
+		if(plus is None):
+			return [num,False]
+		else:
+			if(plus=="+"):
+				return [num,True]
+			else:
+				return [num,False]
+	else:
+		return None
+
+
+
 if (__name__=="__main__"):
-	parser=argparse.ArgumentParser(description='Print DIOGENIX formatted RM frequency statistics')
+	parser=argparse.ArgumentParser(description='Print DIOGENIX formatted RM frequency statistics (exact or exact+ modes)')
 	parser.add_argument('out_tbl',type=str,nargs=1,help="output file path for TSV of sample/RM data table")
 	parser.add_argument('dir_base',type=str,nargs=1,help="base dir for files")
+	parser.add_argument('-exact_plus',type=str,nargs=1,default="0+",help="valid range to report over : may be : 0,0+,1,1+,2,2+, etc.... depending on range")
 	args=parser.parse_args()
 	if(not(args)):
 		parser.print_help()
 	else:
+		exact_val=extractAsItemOrFirstFromList(args.exact_plus)
+		exact_val_validated=validate_filter_stat_arg(exact_val)
+		if(exact_val_validated is None):
+			print "Invalid 'exact_plus' : '"+str(exact_val)+"' "
+			parser.print_help()
+			import sys
+			sys.exit(0)
 		out_tbl=extractAsItemOrFirstFromList(args.out_tbl)
 		dir_base=extractAsItemOrFirstFromList(args.dir_base)
 		if(not(os.path.exists(dir_base))):
 			parser.print_help()
 		else:
-			writeRMInfoToTable(dir_base,out_tbl)
+			writeRMInfoToTable(dir_base,out_tbl,exact_val)
